Put all default implementations of `Age` into `Ages`.
